+++ docs/getting_started/jaeger_example.py
@@ -20,7 +20,7 @@ from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
 trace.set_tracer_provider(TracerProvider())
 
-jaeger_exporter = jaeger.JaegerSpanExporter(
+jaeger_exporter = jaeger.JaegerExporter(
     service_name="my-helloworld-service",
     agent_host_name="localhost",
     agent_port=6831,